+++ src/Services/Expenses/Application/Features/Shared/ExpenseDTO.cs
@@ -5,7 +5,7 @@ public record ExpenseDTO
     public string Id { get; init; } = string.Empty;
     public string Name { get; init; } = string.Empty;
     public string Description { get; init; } = string.Empty;
-    public double Value { get; init; }
+    public decimal Value { get; init; }
     public DateTime Date { get; init; }
     public string CategoryId { get; init; } = string.Empty;
     public string ApplicationUserId { get; init; } = string.Empty;
